feat(chamado-service): adiciona validacao para deletar funcionario

+++ src/main/java/dev/joaorooliveira/help_desk/domain/funcionario/FuncionarioService.java
@@ -1,10 +1,12 @@
 package dev.joaorooliveira.help_desk.domain.funcionario;
 
+import dev.joaorooliveira.help_desk.domain.chamado.ChamadoRepository;
 import dev.joaorooliveira.help_desk.domain.funcionario.dto.FuncionarioAtualizarDTO;
 import dev.joaorooliveira.help_desk.domain.funcionario.dto.FuncionarioFiltroRequestDTO;
 import dev.joaorooliveira.help_desk.domain.funcionario.dto.FuncionarioRequestDTO;
 import dev.joaorooliveira.help_desk.domain.funcionario.dto.FuncionarioResponseDTO;
 import dev.joaorooliveira.help_desk.infra.exception.EntidadeNaoEncontradaException;
+import dev.joaorooliveira.help_desk.infra.exception.RegraNegocioException;
 import dev.joaorooliveira.help_desk.infra.specification.FuncionarioSpecification;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -15,9 +17,11 @@ import org.springframework.transaction.annotation.Transactional;
 public class FuncionarioService {
 
     private final FuncionarioRepository funcionarioRepository;
+    private final ChamadoRepository chamadoRepository;
 
-    public FuncionarioService(FuncionarioRepository funcionarioRepository) {
+    public FuncionarioService(FuncionarioRepository funcionarioRepository, ChamadoRepository chamadoRepository) {
         this.funcionarioRepository = funcionarioRepository;
+        this.chamadoRepository = chamadoRepository;
     }
 
     @Transactional
@@ -38,14 +42,17 @@ public class FuncionarioService {
         return FuncionarioResponseDTO.fromEntity(funcionario);
     }
 
-    //Lembrar de validar se o funcionario nao tem chamados associados antes de deletar
     @Transactional
     public void deletarFuncionario(Long id) {
         if (!funcionarioRepository.existsById(id)) {
             throw new EntidadeNaoEncontradaException("Funcionário não encontrado com o ID: " + id);
         }
+        if(chamadoRepository.existsByFuncionarioId(id)){
+            throw new RegraNegocioException("Não é possível excluir o funcionário, pois existem chamados associados a ele");
+        }
         funcionarioRepository.deleteById(id);
     }
+
 
     @Transactional
     public FuncionarioResponseDTO atualizarFuncionario(Long id , FuncionarioAtualizarDTO funcionarioAtualizarDTO) {
